fix(time): old macos fallback (#13918)
(cherry picked from commit 0e5c3e4822e23088aa3b5ad2488c28740d0b2c3d)
