+++ otherlibs/stdune/src/wait4_stubs.c
@@ -24,8 +24,9 @@ void dune_wait4(value v_pid, value flags) {
 #include <sys/time.h>
 #include <sys/types.h>
 #include <sys/wait.h>
-#include <time.h>
 #include <stdint.h>
+
+#include "dune_clock.h"
 
 #define TAG_WEXITED 0
 #define TAG_WSIGNALED 1
@@ -68,7 +69,7 @@ value dune_wait4(value v_pid, value flags) {
   CAMLlocal2(times, res);
 
   int status, cv_flags;
-  struct timespec tp;
+  int64_t time_ns;
   cv_flags = caml_convert_flag_list(flags, wait_flag_table);
   pid_t pid = Int_val(v_pid);
 
@@ -78,7 +79,7 @@ value dune_wait4(value v_pid, value flags) {
   // returns the pid of the terminated process, or -1 on error
   pid = wait4(pid, &status, cv_flags, &ru);
   int wait_errno = errno;
-  clock_gettime(CLOCK_REALTIME, &tp);
+  time_ns = dune_clock_gettime_ns();
   caml_leave_blocking_section();
   if (pid == 0) {
     CAMLreturn(Val_none);
@@ -107,8 +108,7 @@ value dune_wait4(value v_pid, value flags) {
   res = caml_alloc_tuple(4);
   Store_field(res, 0, Val_int(pid));
   Store_field(res, 1, alloc_process_status(status));
-  Store_field(res, 2,
-              Val_long(((int64_t)tp.tv_sec * 1000000000LL) + (int64_t)tp.tv_nsec));
+  Store_field(res, 2, Val_long(time_ns));
   Store_field(res, 3, times);
   CAMLreturn(caml_alloc_some_compat(res));
 }
